Generate bank accounts controller with only the show method
the show method holds the current users bank account info.
it view will act as an overview. i could have called it `overview` instead,
but show is better because maybe the future there will be a more priveliged
bank_manager that has the right to create, show, update, delete which makes
a lot of sense and is RESTfull.