Add a transfer model, this PORO handels the transaction proccess!
This model plain old ruby model does not inherit from active record
and basically takes the attributes needed to make a transfer
from one user to another and does:

- 1. withdraws the amount from the senders bank_account balance.
- 2. creates a record for the withdrawl event to keep track.
- 3. deposites the amount on the receipient bank_account balance.
- 4. creates a record for the deposit event to keep track.

If any of the steps above fail or halt, it rollesback the database.
All SQL queries executed within the transaction are undone.
Then it raises an exception.

I rescue from that exeption and reload the object in memory because
those are not done automatically by rails. rails has no easy way of
knowing what models were involved.
Once the are reloaded all is in sync. objects in memory and database.

Then i re-raise the exception so that it can bubble up be catched in
the calling layer in our case the controller.
And after that passed i will pass it on to the view layer via to the user
through a notice (flash).

The private methods in this model each has their single responsibility:
- withdraw!
- deposit!
- amount_valid?
- adjust_balance_and_save! (raises an exception if bank_account.balance does not exist)
- create_record! (for our administration and tracking the transfers)
- reload_accounts_and_records (in case of a rollback)

I left the console warnings here to give you an indication of the points i did debugging.
And to sketch my toughts behind this process for you.

I might refactor it into smaller service objects but for now its fine i think for this
demonstration. will do this later in my free time.