Add transfers_controller
In this controller we have 2 main methods.
- new:
provides the transfer form with @users instance variable,
we allready have current_user availeble and we dont want to
transfer to ourself. so we do: where.not
Now we can choose only the other users to transfer to.

- create!:

This method passes the params to the Transfer model transfer class method
and when its done and is saved it redirect the user to the root_path
or we receive an exception from it.

either:
  1. the record is not found.
  2. we did not fill in the amount
  3. any other exeption.
we redirect the user to form and trigger a flash alert.